Add support for a DEFAULT Redis Server
When the “DEFAULT” is chosen, we create the authentication token by
making a request from the server, and we don’t save this token in the
project source (but do provide it to the buildserver in the YAIL code).

If the Redis Server is not DEFAULT, then we persist the Token
property (because it is the password to the non-default redis server)
and we do not over-write it by getting a new one from the server.

Change-Id: I68fa9625772dbccac901d258d4b55e413f356952