Temporarily Remove Off-line storage from CloudDB
Our first version will not have the off-line capability. This is a
temporary situation. After we are “live” we can revert this commit and
complete the off-line version.

Note: Design re-think. Offline capable variables may want to be
separated from “normal” CloudDB variables, or perhaps we only implement
offline capability for list push. The idea being that anyone can push to
a list, even when off-line. When they return to an on-line state, the
push will happen. This is non-destructive if several people are working
off-line, all of their pushes will eventually happen.

Change-Id: I9c9aaa2350e8dc63644e37b1badd15e03eab3b4f