CloudDB Updates
Add a UseSSL property.

Refactor code to use explicit project level variable subscriptions
instead of using the Redis provided variable change events. Instead of
calling Redis “set” we use a LUA script to do both the set and the
publishing of the variable value.

Use EVALSHA instead of EVAL and fallback to EVAL if the server doesn’t
have the script cached. The first use of EVAL will populate the server’s
cache for all users.

Change-Id: I10873bba90aea87fefc01f40f4ba9367d7e7d7a2