Add support for permanent tokens for DEFAULT server
In CloudDB, we normally do not persist the Token for the default
server (it is recomputed each time a project is loaded). We add a
feature that if the token starts with a percent sign (%), then we do
persist the token. At runtime we strip off any leading percent
sign. This permits us to distribute Tokens not associated with an
account and they will be persisted. This permits multiple people to
share a CloudDB variable address space.

Change-Id: I91dd337cccf8970861ef5f075f70af769bc6889b